requirements.txt
accelerate==1.7.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.49.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
azure-cognitiveservices-speech==1.43.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.23
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
Distance==0.1.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
edge-tts==7.0.0
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.103.1
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.18.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Session==0.8.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
g2p-en==2.1.0
gradio==5.22.0
gradio_client==1.8.0
groovy==0.1.2
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
identify==2.6.9
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jieba==0.39
jieba-fast==0.53
Jinja2==3.1.2
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
langdetect==1.0.9
lightning-utilities==0.14.2
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
modelscope==1.24.0
more-itertools==10.6.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
onnxruntime==1.21.0
open-llm-vtuber @ file:///F:/Open-LLM-VTuber
openai==1.65.5
OpenCC==1.1.9
openpyxl==3.1.2
orjson==3.10.15
packaging==24.2
pandas==2.2.2
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.3.7
portalocker==3.1.1
pre_commit==4.2.0
proces==0.1.7
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.4.8
pycorrector==0.3.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.8.0
pyparsing==3.2.3
pypinyin==0.53.0
pypiwin32==223
pyreadline3==3.5.4
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
pytorch-lightning==2.5.1
pyttsx3==2.98
pytz==2025.1
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.2
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
sherpa-onnx==1.11.2
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
srt==3.5.3
starlette==0.27.0
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0+cu118
torchaudio==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.9
uvicorn==0.23.2
virtualenv==20.29.3
watchfiles==1.0.4
websocket-client==1.8.0
websockets==13.1
Werkzeug==2.3.7
win32_setctime==1.2.0
wordsegment==1.3.1
xlrd==2.0.1
xxhash==3.5.0
yarl==1.18.3
zhipuai==2.1.5.20250421
